Split peer locking tests
