+++ dev/com.ibm.ws.transaction.cloud_fat.peerlocking.1/publish/servers/com.ibm.ws.transaction_LKCLOUD002/server.xml
