+++ dev/com.ibm.ws.transaction.cloud_fat.peerlocking.1/publish/servers/defaultAttributesServer1/bootstrap.properties
