+++ dev/com.ibm.ws.transaction.cloud_fat.peerlocking.1/publish/servers/longPeerStaleTimeServer2/bootstrap.properties
