+++ dev/com.ibm.ws.transaction.cloud_fat.peerlocking.1/publish/servers/peerLockingDisabledServer1/jvm.options
