+++ dev/com.ibm.ws.transaction.cloud_fat.peerlocking.2/.gitignore
@@ -1,0 +1,1 @@
+/publish
