+++ dev/com.ibm.ws.transaction.cloud_fat.peerlocking.2/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>com.ibm.ws.transaction.cloud_fat.peerlocking</name>
+	<name>com.ibm.ws.transaction.cloud_fat.peerlocking.2</name>
 	<comment></comment>
 	<projects>
 	</projects>
