+++ dev/com.ibm.ws.transaction.cloud_fat.peerlocking.2/.settings/bndtools.core.prefs
@@ -1,0 +1,2 @@
+compileErrorAction=build
+eclipse.preferences.version=1
