+++ dev/com.ibm.ws.transaction.cloud_fat.peerlocking.2/build.gradle
@@ -1,0 +1,58 @@
+/*******************************************************************************
+ * Copyright (c) 2023 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+
+// Define G:A:V coordinates of each dependency
+dependencies {
+  requiredLibs project(":com.ibm.ws.tx.embeddable"),
+               project(":com.ibm.ws.transaction.cloud_fat.base"),
+               project(':com.ibm.ws.transaction.cloud_fat.peerlocking.1'),
+   			   project(":com.ibm.ws.transaction.fat.util"),
+               project(':io.openliberty.org.apache.commons.logging')
+}
+
+File sharedDir = new File(autoFvtDir, 'publish/shared/resources')
+
+task addDerbyToSharedDir(type: Copy) {
+  from configurations.derby
+  into new File(sharedDir, 'derby')
+  rename 'derby-.*.jar', 'derby.jar'
+}
+
+task copyCommonFiles {
+  doLast {
+    /*
+     * Copy common publish files.
+     */
+    copy {
+      from project(':com.ibm.ws.transaction.cloud_fat.peerlocking.1').file('publish')
+      into new File('publish')
+    }
+  }
+}
+
+addRequiredLibraries.dependsOn copyCommonFiles
+addRequiredLibraries.dependsOn addDerbyToSharedDir
+addRequiredLibraries.dependsOn addJakartaTransformer
+
+jar {
+  dependsOn copyCommonFiles
+}
+
+autoFVT.dependsOn copyCommonFiles
+
+clean.doLast {
+  if (file('publish').exists()){
+    file('publish').deleteDir()
+  }
+  if (file('build').exists()){
+    file('build').deleteDir()
+  }
+}
