--- dev/com.ibm.ws.transaction.cloud_fat.peerlocking/fat/src/com/ibm/ws/transaction/test/.gitignore
@@ -1,1 +1,0 @@
-/tests
